home

+++ src/SSCMS.Core/Repositories/LogRepository.cs
@@ -12,18 +12,20 @@ using SSCMS.Utils;
 
 namespace SSCMS.Core.Repositories
 {
-    public partial class LogRepository : ILogRepository
+    public class LogRepository : ILogRepository
     {
         private readonly Repository<Log> _repository;
         private readonly IConfigRepository _configRepository;
         private readonly IAdministratorRepository _administratorRepository;
+        private readonly IUserRepository _userRepository;
         private readonly IErrorLogRepository _errorLogRepository;
 
-        public LogRepository(ISettingsManager settingsManager, IConfigRepository configRepository, IAdministratorRepository administratorRepository, IErrorLogRepository errorLogRepository)
+        public LogRepository(ISettingsManager settingsManager, IConfigRepository configRepository, IAdministratorRepository administratorRepository, IUserRepository userRepository, IErrorLogRepository errorLogRepository)
         {
             _repository = new Repository<Log>(settingsManager.Database, settingsManager.Redis);
             _configRepository = configRepository;
             _administratorRepository = administratorRepository;
+            _userRepository = userRepository;
             _errorLogRepository = errorLogRepository;
         }
 
@@ -33,9 +35,78 @@ namespace SSCMS.Core.Repositories
 
         public List<TableColumn> TableColumns => _repository.TableColumns;
 
-        public async Task InsertAsync(Log log)
+        public async Task AddAdminLogAsync(Administrator admin, string action, string summary = "")
         {
-            await _repository.InsertAsync(log);
+            var config = await _configRepository.GetAsync();
+            if (!config.IsLogAdmin) return;
+
+            try
+            {
+                await DeleteIfThresholdAsync();
+
+                if (!string.IsNullOrEmpty(action))
+                {
+                    action = StringUtils.MaxLengthText(action, 250);
+                }
+                if (!string.IsNullOrEmpty(summary))
+                {
+                    summary = StringUtils.MaxLengthText(summary, 250);
+                }
+
+                var log = new Log
+                {
+                    Id = 0,
+                    AdminId = admin.Id,
+                    IpAddress = string.Empty,
+                    Action = action,
+                    Summary = summary
+                };
+
+                await _repository.InsertAsync(log);
+
+                await _administratorRepository.UpdateLastActivityDateAsync(admin);
+            }
+            catch (Exception ex)
+            {
+                await _errorLogRepository.AddErrorLogAsync(ex);
+            }
+        }
+
+        public async Task AddUserLogAsync(User user, string action, string summary = "")
+        {
+            var config = await _configRepository.GetAsync();
+            if (!config.IsLogAdmin) return;
+
+            try
+            {
+                await DeleteIfThresholdAsync();
+
+                if (!string.IsNullOrEmpty(action))
+                {
+                    action = StringUtils.MaxLengthText(action, 250);
+                }
+                if (!string.IsNullOrEmpty(summary))
+                {
+                    summary = StringUtils.MaxLengthText(summary, 250);
+                }
+
+                var log = new Log
+                {
+                    Id = 0,
+                    AdminId = user.Id,
+                    IpAddress = string.Empty,
+                    Action = action,
+                    Summary = summary
+                };
+
+                await _repository.InsertAsync(log);
+
+                await _userRepository.UpdateLastActivityDateAsync(user);
+            }
+            catch (Exception ex)
+            {
+                await _errorLogRepository.AddErrorLogAsync(ex);
+            }
         }
 
         public async Task DeleteIfThresholdAsync()
@@ -51,14 +122,21 @@ namespace SSCMS.Core.Repositories
             );
         }
 
-        public async Task DeleteAllAsync()
+        public async Task DeleteAllAdminLogsAsync()
         {
-            await _repository.DeleteAsync();
+            await _repository.DeleteAsync(Q.Where(nameof(Log.AdminId), ">", 0));
         }
 
-        private Query GetQuery(int adminId, string keyword, string dateFrom, string dateTo)
+        public async Task DeleteAllUserLogsAsync()
         {
-            var query = Q.OrderByDesc(nameof(Log.Id));
+            await _repository.DeleteAsync(Q.Where(nameof(Log.UserId), ">", 0));
+        }
+
+        private Query GetAdminQuery(int adminId, string keyword, string dateFrom, string dateTo)
+        {
+            var query = Q
+                .Where(nameof(Log.AdminId), ">", 0)
+                .OrderByDesc(nameof(Log.Id));
 
             if (adminId == 0 && string.IsNullOrEmpty(keyword) && string.IsNullOrEmpty(dateFrom) && string.IsNullOrEmpty(dateTo))
             {
@@ -80,24 +158,72 @@ namespace SSCMS.Core.Repositories
 
             if (!string.IsNullOrEmpty(dateFrom))
             {
-                query.WhereDate(nameof(Log.AddDate), ">=", TranslateUtils.ToDateTime(dateFrom));
+                query.WhereDate(nameof(Log.CreatedDate), ">=", TranslateUtils.ToDateTime(dateFrom));
             }
             if (!string.IsNullOrEmpty(dateTo))
             {
-                query.WhereDate(nameof(Log.AddDate), "<=", TranslateUtils.ToDateTime(dateTo));
+                query.WhereDate(nameof(Log.CreatedDate), "<=", TranslateUtils.ToDateTime(dateTo));
             }
 
             return query;
         }
 
-        public async Task<int> GetCountAsync(int adminId, string keyword, string dateFrom, string dateTo)
+        public async Task<int> GetAdminLogsCountAsync(int adminId, string keyword, string dateFrom, string dateTo)
         {
-            return await _repository.CountAsync(GetQuery(adminId, keyword, dateFrom, dateTo));
+            return await _repository.CountAsync(GetAdminQuery(adminId, keyword, dateFrom, dateTo));
         }
 
-        public async Task<List<Log>> GetAllAsync(int adminId, string keyword, string dateFrom, string dateTo, int offset, int limit)
+        public async Task<List<Log>> GetAdminLogsAsync(int adminId, string keyword, string dateFrom, string dateTo, int offset, int limit)
         {
-            var query = GetQuery(adminId, keyword, dateFrom, dateTo);
+            var query = GetAdminQuery(adminId, keyword, dateFrom, dateTo);
+            query.Offset(offset).Limit(limit);
+            return await _repository.GetAllAsync(query);
+        }
+
+        private Query GetUserQuery(int userId, string keyword, string dateFrom, string dateTo)
+        {
+            var query = Q
+                .Where(nameof(Log.UserId), ">", 0)
+                .OrderByDesc(nameof(Log.Id));
+
+            if (userId == 0 && string.IsNullOrEmpty(keyword) && string.IsNullOrEmpty(dateFrom) && string.IsNullOrEmpty(dateTo))
+            {
+                return query;
+            }
+
+            if (userId > 0)
+            {
+                query.Where(nameof(Log.UserId), userId);
+            }
+
+            if (!string.IsNullOrEmpty(keyword))
+            {
+                var like = $"%{keyword}%";
+                query.Where(q =>
+                    q.WhereLike(nameof(Log.Action), like).OrWhereLike(nameof(Log.Summary), like)
+                );
+            }
+
+            if (!string.IsNullOrEmpty(dateFrom))
+            {
+                query.WhereDate(nameof(Log.CreatedDate), ">=", TranslateUtils.ToDateTime(dateFrom));
+            }
+            if (!string.IsNullOrEmpty(dateTo))
+            {
+                query.WhereDate(nameof(Log.CreatedDate), "<=", TranslateUtils.ToDateTime(dateTo));
+            }
+
+            return query;
+        }
+
+        public async Task<int> GetUserLogsCountAsync(int userId, string keyword, string dateFrom, string dateTo)
+        {
+            return await _repository.CountAsync(GetUserQuery(userId, keyword, dateFrom, dateTo));
+        }
+
+        public async Task<List<Log>> GetUserLogsAsync(int userId, string keyword, string dateFrom, string dateTo, int offset, int limit)
+        {
+            var query = GetUserQuery(userId, keyword, dateFrom, dateTo);
             query.Offset(offset).Limit(limit);
             return await _repository.GetAllAsync(query);
         }
@@ -198,16 +324,16 @@ SELECT COUNT(*) AS AddNum, AddYear FROM (
             var dict = new Dictionary<string, int>();
 
             var query = Q
-                .WhereDate(nameof(Log.AddDate), ">", DateTime.Now.AddDays(-30))
+                .WhereDate(nameof(Log.CreatedDate), ">", DateTime.Now.AddDays(-30))
                 .Where(nameof(Log.AdminId), ">", 0);
 
             if (dateFrom > Constants.SqlMinValue)
             {
-                query.WhereDate(nameof(Log.AddDate), ">=", dateFrom);
+                query.WhereDate(nameof(Log.CreatedDate), ">=", dateFrom);
             }
             if (dateTo != Constants.SqlMinValue)
             {
-                query.WhereDate(nameof(Log.AddDate), "<", dateTo);
+                query.WhereDate(nameof(Log.CreatedDate), "<", dateTo);
             }
 
             var list = await _repository.GetAllAsync<(int adminId, int count)>(query
@@ -253,5 +379,146 @@ SELECT COUNT(*) AS AddNum, AddYear FROM (
             
             return dict;
         }
+
+        public async Task<List<Log>> GetUserLogsAsync(int userId, int offset, int limit)
+        {
+            return await _repository.GetAllAsync(Q
+                .Where(nameof(Log.UserId), userId)
+                .Offset(offset)
+                .Limit(limit)
+                .OrderByDesc(nameof(Log.Id))
+            );
+        }
+
+        public async Task<List<Log>> GetAdminLogsAsync(int adminId, int offset, int limit)
+        {
+            return await _repository.GetAllAsync(Q
+                .Where(nameof(Log.AdminId), adminId)
+                .Offset(offset)
+                .Limit(limit)
+                .OrderByDesc(nameof(Log.Id))
+            );
+        }
     }
 }
+
+
+//using System;
+//using System.Collections.Generic;
+//using System.Threading.Tasks;
+//using Datory;
+//using SqlKata;
+//using SSCMS.Models;
+//using SSCMS.Repositories;
+//using SSCMS.Utils;
+
+//namespace SSCMS.Core.Repositories
+//{
+//    public partial class UserLogRepository : IUserLogRepository
+//    {
+//        private readonly Repository<UserLog> _repository;
+//        private readonly IConfigRepository _configRepository;
+//        private readonly IErrorLogRepository _errorLogRepository;
+
+//        public UserLogRepository(ISettingsManager settingsManager, IConfigRepository configRepository, IErrorLogRepository errorLogRepository)
+//        {
+//            _repository = new Repository<UserLog>(settingsManager.Database, settingsManager.Redis);
+//            _configRepository = configRepository;
+//            _errorLogRepository = errorLogRepository;
+//        }
+
+//        public IDatabase Database => _repository.Database;
+
+//        public string TableName => _repository.TableName;
+
+//        public List<TableColumn> TableColumns => _repository.TableColumns;
+
+//        public async Task InsertAsync(UserLog userLog)
+//        {
+//            await _repository.InsertAsync(userLog);
+//        }
+
+//        public async Task DeleteIfThresholdAsync()
+//        {
+//            var config = await _configRepository.GetAsync();
+//            if (!config.IsTimeThreshold) return;
+
+//            var days = config.TimeThreshold;
+//            if (days <= 0) return;
+
+//            await _repository.DeleteAsync(Q
+//                .Where(nameof(UserLog.CreatedDate), "<", DateTime.Now.AddDays(-days))
+//            );
+//        }
+
+//        public async Task DeleteAllAsync()
+//        {
+//            await _repository.DeleteAsync();
+//        }
+
+//        private Query GetQuery(int userId, string keyword, string dateFrom, string dateTo)
+//        {
+//            var query = Q.OrderByDesc(nameof(UserLog.Id));
+
+//            if (userId == 0 && string.IsNullOrEmpty(keyword) && string.IsNullOrEmpty(dateFrom) && string.IsNullOrEmpty(dateTo))
+//            {
+//                return query;
+//            }
+
+//            if (userId > 0)
+//            {
+//                query.Where(nameof(UserLog.UserId), userId);
+//            }
+
+//            if (!string.IsNullOrEmpty(keyword))
+//            {
+//                var like = $"%{keyword}%";
+//                query.Where(q =>
+//                    q.WhereLike(nameof(UserLog.Action), like).OrWhereLike(nameof(UserLog.Summary), like)
+//                );
+//            }
+
+//            if (!string.IsNullOrEmpty(dateFrom))
+//            {
+//                query.WhereDate(nameof(UserLog.CreatedDate), ">=", TranslateUtils.ToDateTime(dateFrom));
+//            }
+//            if (!string.IsNullOrEmpty(dateTo))
+//            {
+//                query.WhereDate(nameof(UserLog.CreatedDate), "<=", TranslateUtils.ToDateTime(dateTo));
+//            }
+
+//            return query;
+//        }
+
+//        public async Task<int> GetCountAsync(int userId, string keyword, string dateFrom, string dateTo)
+//        {
+//            return await _repository.CountAsync(GetQuery(userId, keyword, dateFrom, dateTo));
+//        }
+
+//        public async Task<List<UserLog>> GetAllAsync(int userId, string keyword, string dateFrom, string dateTo, int offset, int limit)
+//        {
+//            var query = GetQuery(userId, keyword, dateFrom, dateTo);
+//            query.Offset(offset).Limit(limit);
+//            return await _repository.GetAllAsync(query);
+//        }
+
+//        public async Task<List<UserLog>> GetLogsAsync(int userId, int offset, int limit)
+//        {
+//            return await _repository.GetAllAsync(Q
+//                .Where(nameof(UserLog.UserId), userId)
+//                .Offset(offset)
+//                .Limit(limit)
+//                .OrderByDesc(nameof(UserLog.Id))
+//            );
+//        }
+
+//        public async Task<UserLog> InsertAsync(int userId, UserLog log)
+//        {
+//            log.UserId = userId;
+
+//            log.Id = await _repository.InsertAsync(log);
+
+//            return log;
+//        }
+//    }
+//}
